[major] Migration to @babel/* packages
