+++ packages/bundle/package.json
@@ -7,10 +7,10 @@
   "module": "./index",
   "react-native": "./index",
   "scripts": {
-    "test:web": "mocha --colors --require babel-register test/*.test.js",
+    "test:web": "mocha --colors --require setup-env test/*.test.js",
     "test": "npm run build && nyc --reporter=text --reporter=lcov npm run test:web",
-    "prepublish": "npm run build",
-    "build": "rm -rf ./lib && babel ./index.js ./walker/*.js ./dimensions/*.js -d ./lib",
+    "prepublishOnly": "npm run build",
+    "build": "rm -rf ./lib && babel ./index.js -d ./lib && babel ./walker/*.js -d ./lib/walker && babel ./dimensions/*.js -d ./lib/dimensions",
     "lint": "eslint-godaddy *.js ./test/*.js"
   },
   "dependencies": {
@@ -23,28 +23,37 @@
     "svgo": "^1.0.3"
   },
   "devDependencies": {
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
     "assume": "^1.5.2",
-    "puppeteer": "^1.1.0",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.26.0",
     "eslint": "^4.14.0",
     "eslint-config-godaddy": "^2.1.0",
     "eslint-plugin-json": "^1.2.0",
     "eslint-plugin-mocha": "^4.11.0",
     "eslint-plugin-react": "^7.5.1",
     "mocha": "^5.0.1",
-    "nyc": "^11.4.1"
+    "nyc": "^11.4.1",
+    "puppeteer": "^1.1.0",
+    "setup-env": "^1.2.1"
   },
   "babel": {
     "plugins": [
-      "transform-object-rest-spread"
+      "@babel/plugin-proposal-object-rest-spread"
     ],
     "presets": [
-      "babel-preset-es2015",
-      "babel-preset-react"
+      [
+        "@babel/preset-env",
+        {
+          "targets": {
+            "node": 9
+          }
+        }
+      ],
+      "@babel/preset-react"
     ]
   },
   "browserify": {
