+++ packages/list/index.js
@@ -1,9 +1,9 @@
 const debug = require('diagnostics')('asset:list');
 const table = require('markdown-table');
 const EventEmitter = require('events');
+const mkdirp = require('mkdirp');
 const bytes = require('bytes');
 const path = require('path');
-const mkdirp = require('mkdirp');
 const zlib = require('zlib');
 const url = require('url');
 const fs = require('fs');
