+++ packages/list/package.json
@@ -7,7 +7,7 @@
   "module": "./index",
   "react-native": "./index",
   "scripts": {
-    "test:web": "mocha --colors --require babel-register test/*.test.js",
+    "test:web": "mocha --colors --require setup-env test/*.test.js",
     "test": "nyc --reporter=text --reporter=lcov npm run test:web",
     "prepublishOnly": "npm run build",
     "build": "rm -rf ./lib && babel ./index.js -d ./lib",
@@ -20,25 +20,27 @@
     "mkdirp": "^0.5.1"
   },
   "devDependencies": {
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/register": "^7.0.0",
     "assume": "^1.5.2",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-register": "^6.26.0",
     "clone": "^2.1.1",
     "eslint": "^4.14.0",
     "eslint-config-godaddy": "^2.1.0",
     "eslint-plugin-json": "^1.2.0",
     "eslint-plugin-mocha": "^4.11.0",
     "mocha": "^4.1.0",
-    "nyc": "^11.4.1"
+    "nyc": "^11.4.1",
+    "setup-env": "^1.2.1"
   },
   "babel": {
     "plugins": [
-      "transform-object-rest-spread"
+      "@babel/plugin-proposal-object-rest-spread"
     ],
     "presets": [
-      "babel-preset-es2015"
+      "@babel/preset-env"
     ]
   },
   "browserify": {
