+++ packages/parser/package.json
@@ -7,10 +7,10 @@
   "module": "./index",
   "react-native": "./index",
   "scripts": {
-    "test:web": "mocha --colors --require babel-register --require test/setup.js test/*.test.js",
+    "test:web": "mocha --colors --require setup-env test/*.test.js",
     "test": "nyc --reporter=text --reporter=lcov npm run test:web",
-    "prepublish": "npm run build",
-    "build": "rm -rf ./lib && babel ./asset.js ./index.js ./packer/*.js ./passthrough/*.js ./transform.js -d ./lib",
+    "prepublishOnly": "npm run build",
+    "build": "rm -rf ./lib && babel ./asset.js ./index.js ./packer.js ./passthrough.js ./transform.js -d ./lib",
     "lint": "eslint-godaddy-react *.js packer/*.js passthrough/*.js ./test/*.js"
   },
   "homepage": "https://github.com/godaddy/asset-system/tree/master/packages/parser",
@@ -25,11 +25,11 @@
   ],
   "babel": {
     "plugins": [
-      "transform-object-rest-spread"
+      "@babel/plugin-proposal-object-rest-spread"
     ],
     "presets": [
-      "babel-preset-es2015",
-      "babel-preset-react"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ]
   },
   "browserify": {
@@ -43,13 +43,14 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
     "assume": "^1.5.2",
     "assume-enzyme": "^0.0.1",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.26.0",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.14.0",
@@ -63,11 +64,12 @@
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-test-renderer": "^16.2.0"
+    "react-test-renderer": "^16.2.0",
+    "setup-env": "^1.2.1"
   },
   "dependencies": {
     "diagnostics": "^1.1.0",
     "rip-out": "^1.0.0",
-    "svgs": "^3.2.0"
+    "svgs": "^4.0.0"
   }
 }
