+++ packages/provider/package.json
@@ -7,10 +7,10 @@
   "module": "./index",
   "react-native": "./index",
   "scripts": {
-    "test:web": "mocha --colors --require babel-register --require test/setup.js test/*.test.js",
+    "test:web": "mocha --colors --require setup-env --require test/nock.js test/*.test.js",
     "test": "nyc --reporter=text --reporter=lcov npm run test:web",
     "prepublish": "npm run build",
-    "build": "rm -rf ./lib && babel ./asset/*.js ./fallback/*.js ./index.js ./loader/*.js ./provider/*.js ./wrapper/*.js -d ./lib",
+    "build": "rm -rf ./lib && babel ./index.js -d ./lib && babel ./asset/*.js -d ./lib/asset && babel ./fallback/*.js -d ./lib/fallback && babel ./loader/*.js -d ./lib/loader && babel ./provider/*.js -d ./lib/provider && babel ./wrapper/*.js -d ./lib/wrapper",
     "lint": "eslint-godaddy-react *.js ./provider/*.js ./fallback/*.js ./asset/*.js ./test/*.js"
   },
   "homepage": "https://github.com/godaddy/asset-system/tree/master/packages/provider",
@@ -28,11 +28,11 @@
   ],
   "babel": {
     "plugins": [
-      "transform-object-rest-spread"
+      "@babel/plugin-proposal-object-rest-spread"
     ],
     "presets": [
-      "babel-preset-es2015",
-      "babel-preset-react"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ]
   },
   "browserify": {
@@ -46,12 +46,13 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
     "assume": "^1.5.2",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.26.0",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.14.0",
@@ -67,7 +68,8 @@
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-test-renderer": "^16.2.0"
+    "react-test-renderer": "^16.2.0",
+    "setup-env": "^1.2.1"
   },
   "dependencies": {
     "asset-parser": "^2.1.0",
