+++ packages/test/CHANGELOG.md
@@ -1,0 +1,13 @@
+# CHANGELOG
+
+### 2.1.0
+
+- Point to ES5 build for main/browser packages.
+
+### 2.0.0
+
+- Removed React 15 support.
+
+### 1.0.0
+
+- Initial public release.
