+++ packages/test/README.md
@@ -1,0 +1,12 @@
+# asset-test
+
+This is not the module you are looking for, this simply holds our test fixtures
+so they can be easily shared between all packages in this mono repo.
+
+## Installation
+
+Nope, not gonna tell you, because you don't need this package.
+
+## License
+
+[MIT](LICENSE)
