+++ packages/webpack/package.json
@@ -7,7 +7,7 @@
   "module": "./index",
   "react-native": "./index",
   "scripts": {
-    "test:web": "mocha --colors --require babel-register test/*.test.js",
+    "test:web": "mocha --colors --require setup-env test/*.test.js",
     "test": "npm run build && nyc --reporter=text --reporter=lcov npm run test:web",
     "prepublish": "npm run build",
     "build": "babel ./index.js ./loader.js -d ./lib",
@@ -25,11 +25,11 @@
   "license": "MIT",
   "babel": {
     "plugins": [
-      "transform-object-rest-spread"
+      "@babel/plugin-proposal-object-rest-spread"
     ],
     "presets": [
-      "babel-preset-es2015",
-      "babel-preset-react"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ]
   },
   "browserify": {
@@ -43,12 +43,12 @@
     ]
   },
   "devDependencies": {
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
     "assume": "^1.5.2",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.26.0",
     "eslint": "^4.14.0",
     "eslint-config-godaddy": "^2.1.0",
     "eslint-plugin-json": "^1.2.0",
@@ -56,6 +56,7 @@
     "eslint-plugin-react": "^7.5.1",
     "mocha": "^4.1.0",
     "nyc": "^11.4.1",
+    "setup-env": "^1.2.1",
     "webpack": "^3.10.0"
   },
   "dependencies": {
